requirements.txt
asgiref==3.4.1
backcall==0.1.0
beautifulsoup4==4.9.3
boto3==1.20.26
botocore==1.23.26
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==2.0.9
dataclasses-json==0.4.2
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.10
django-cors-headers==3.10.1
django-heroku==0.3.1
django-user-agents==0.4.0
djangorestframework==3.12.4
dynamic-rest==2.1.1
future==0.18.2
gunicorn==20.0.4
hashids==1.3.1
idna==2.9
importlib-metadata==4.8.3
inflection==0.5.1
ipython==7.14.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.17.0
jmespath==0.10.0
Markdown==3.3.6
marshmallow==3.5.2
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.10.0
python-dateutil==2.8.2
python-youtube==0.6.0
pytz==2020.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.10.14
s3transfer==0.5.0
sentry-sdk==0.19.5
six==1.14.0
soupsieve==2.1
sqlparse==0.3.1
static3==0.7.0
stringcase==1.2.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.2
typing-inspect==0.6.0
ua-parser==0.10.0
urllib3==1.26.7
user-agents==2.1
wcwidth==0.1.9
whitenoise==5.0.1
zipp==3.6.0
